+++ dns/const_test.go
@@ -3,7 +3,7 @@ package dns
 import "testing"
 
 func TestTypeToString(t *testing.T) {
-	for _, v := range []int{
+	for _, v := range []uint16{
 		TypeA,
 		TypePTR,
 		TypeTXT,
